back out the mapper package move ... will deal with that separately
